Convert shared components to typescript
Also move to the common directory
